Migrate to using elementId
The cypher function `id` has been deprecated since Neo4j 5.0.
This PR adds the capability of sniffing the DBMS version so that the DiffService
can write an update query that matches the server version.

Likewise has `entity.id()` in the driver been deprecated. The driver will set
`elementId` to `id.toString()` if the server does not provide it (4.4 and
before). Therefore, it's safe to always use `entity.elementId()` regardless of
the server version.